DjangoMongoSRL/requirements.txt
asgiref==3.8.1
bson==0.5.8
dataclasses==0.6
Django==3.0.5
djangorestframework==3.12.4
djongo==1.3.3
dnspython==2.6.1
pymongo==3.12.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.12.2
drf-nested-routers==0.92.5
django-filter==2.4.0
